fixed read function not being called
