+++ CMakeLists.txt
@@ -18,7 +18,8 @@ set(SRC_FILES
     ${CMAKE_CURRENT_SOURCE_DIR}/src/utils/i2c.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/src/utils/i2c.hpp
     ${CMAKE_CURRENT_SOURCE_DIR}/src/utils/server.cpp
-)
+        ${CMAKE_CURRENT_SOURCE_DIR}/src/utils/mq3_2.cpp
+        )
 # Set up test files
 set(TEST_FILES
     ${CMAKE_CURRENT_SOURCE_DIR}/test/buffer_test.cpp
