+++ src/main.cpp
@@ -1,6 +1,7 @@
 #include "utils/server.h"
 #include <spdlog/spdlog.h>
 #include <string>
+#include "mq3sensor.h"
 class HelloResp : public serverCallback {
   void serverAction() { spdlog::info("Running Hello Resp"); }
 };
@@ -12,12 +13,12 @@ class DateResp : public serverCallback {
 int main() {
   spdlog::info("Hello Dev");
 
-  HelloResp hello;
-  DateResp date;
-  Server server = Server();
+    MQ3Sensor mq3_sensor(11, 9, 10, 8, 4);
 
-  server.register_callback_action("/hello", &hello);
-  server.register_callback_action("/date", &date);
-  server.run();
-  return 0;
+    while (true) {
+        float ppm = mq3_sensor.get_sensor_ppm();
+        std::cout << "Sensor PPM: " << ppm << std::endl;
+    }
+
+    return 0;
 }
